+++ react-app/src/components/map/parcelPopup/parcelPopup.css
